// src/front/src/api/datasource/models/pickerDataItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX DataSource API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface PickerDataItem {
	/** @nullable */
	gid?: string | null;
	recordId?: number;
	/** @nullable */
	label?: string | null;
	/** @nullable */
	subLabel?: string | null;
}